Navigation  menu for desktop removed from the Landing Page
