+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Box } from "@mui/material"
+import { Box, Paper, Stack } from "@mui/material"
 import Container from "@mui/material/Container"
 import { a, useTransition } from "@react-spring/web"
 import React, { useEffect } from "react"
@@ -7,6 +7,8 @@ import { Navigate, Route, Routes, useLocation } from "react-router-dom"
 import { AccountIndexRoute } from "./AccountIndexRoute"
 import { ErrorBoundary } from "./components/ErrorBoundary"
 import { Header } from "./components/Header/Header"
+import { MenuDrawerContents } from "./components/Header/MenuDrawerContents"
+import { APP_VERSION, GIT_HASH } from "./env"
 import FourZeroFourPage from "./pages/404"
 import AssetPage from "./pages/AssetPage/AssetPage"
 import AssetsPage from "./pages/AssetsPage/AssetsPage"
@@ -19,6 +21,10 @@ import { PnLPage } from "./pages/PnLPage/PnLPage"
 import TransactionsPage from "./pages/TransactionsPage/TransactionsPage"
 import { $pendingTask } from "./stores/task-store"
 import { SPRING_CONFIGS } from "./utils/utils"
+
+function toggleOpen() {
+  // intentionally left blank
+}
 
 export default function App() {
   const location = useLocation()
@@ -55,87 +61,109 @@ export default function App() {
   }, [])
 
   return (
-    <Box sx={{ flex: 1 }}>
-      <Box
-        sx={{
-          "html[data-mui-color-scheme='dark'] &": {
-            background: `radial-gradient(
+    <Stack direction="row">
+      {pathname !== "/" ? (
+        <Paper
+          sx={{
+            display: { xl: "inline-flex", xs: "none" },
+            height: "100vh",
+            minWidth: 300,
+          }}
+          elevation={1}
+        >
+          <MenuDrawerContents
+            open={true}
+            toggleOpen={toggleOpen}
+            appVer={APP_VERSION}
+            gitHash={GIT_HASH}
+          />
+        </Paper>
+      ) : null}
+
+      <Box sx={{ flex: 1 }}>
+        <Box
+          sx={{
+            "html[data-mui-color-scheme='dark'] &": {
+              background: `radial-gradient(
               max(50vw, 400px) min(500px, 50vw) at 50% 0px,
               rgba(var(--mui-palette-accent-mainChannel) / 0.25), #fff0
             )`,
-            height: "100%",
-            left: 0,
-            pointerEvents: "none",
-            position: "absolute",
-            right: 0,
-            top: 0,
-            width: "100%",
-            zIndex: -1,
-          },
-        }}
-      />
-      {pathname !== "/" ? <Header /> : <LandingPageHeader />}
-      <Container disableGutters maxWidth="xl" sx={{ paddingTop: 2, paddingX: { xs: 2 } }}>
-        {transitions((styles, item, { key }) => (
-          <a.div
-            style={
-              {
-                ...styles,
-                maxWidth: 1536 - 32,
-                paddingBottom: 24,
-                position: "absolute",
-                width: "calc(100% - 32px)",
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-              } as any
-            }
-          >
-            <ErrorBoundary>
-              <Routes location={item}>
-                <Route path="/u/:accountIndex" element={<AccountIndexRoute />}>
+              height: "100%",
+              left: 0,
+              pointerEvents: "none",
+              position: "absolute",
+              right: 0,
+              top: 0,
+              width: "100%",
+              zIndex: -1,
+            },
+          }}
+        />
+        {pathname !== "/" ? <Header /> : <LandingPageHeader />}
+        <Container disableGutters maxWidth="xl" sx={{ paddingTop: 2, paddingX: { xs: 2 } }}>
+          {transitions((styles, item, { key }) => (
+            <a.div
+              style={
+                {
+                  ...styles,
+                  maxWidth: 1536 - 32,
+                  paddingBottom: 24,
+                  position: "absolute",
+                  width: "calc(100% - 32px)",
+                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                } as any
+              }
+            >
+              <ErrorBoundary>
+                <Routes location={item}>
+                  <Route path="/u/:accountIndex" element={<AccountIndexRoute />}>
+                    <Route
+                      index
+                      element={<BalancesPage show={key === pathname && appPath === ""} />}
+                    />
+                    <Route
+                      path="pnl"
+                      element={<PnLPage show={key === pathname && appPath === "pnl"} />}
+                    />
+                    <Route
+                      path="asset/:assetId"
+                      element={<AssetPage show={key === pathname && appPath.includes("asset")} />}
+                    />
+                    <Route
+                      path="import-data"
+                      element={
+                        <ImportDataPage show={key === pathname && appPath === "import-data"} />
+                      }
+                    />
+                    <Route
+                      path="transactions"
+                      element={
+                        <TransactionsPage show={key === pathname && appPath === "transactions"} />
+                      }
+                    />
+                    <Route
+                      path="audit-logs"
+                      element={
+                        <AuditLogsPage show={key === pathname && appPath === "audit-logs"} />
+                      }
+                    />
+                    <Route
+                      path="assets"
+                      element={<AssetsPage show={key === pathname && appPath === "assets"} />}
+                    />
+                    <Route path="*" element={<FourZeroFourPage show />} />
+                  </Route>
                   <Route
-                    index
-                    element={<BalancesPage show={key === pathname && appPath === ""} />}
+                    path="/"
+                    element={<LandingPage show={key === pathname && pathname === "/"} />}
                   />
-                  <Route
-                    path="pnl"
-                    element={<PnLPage show={key === pathname && appPath === "pnl"} />}
-                  />
-                  <Route
-                    path="asset/:assetId"
-                    element={<AssetPage show={key === pathname && appPath.includes("asset")} />}
-                  />
-                  <Route
-                    path="import-data"
-                    element={
-                      <ImportDataPage show={key === pathname && appPath === "import-data"} />
-                    }
-                  />
-                  <Route
-                    path="transactions"
-                    element={
-                      <TransactionsPage show={key === pathname && appPath === "transactions"} />
-                    }
-                  />
-                  <Route
-                    path="audit-logs"
-                    element={<AuditLogsPage show={key === pathname && appPath === "audit-logs"} />}
-                  />
-                  <Route
-                    path="assets"
-                    element={<AssetsPage show={key === pathname && appPath === "assets"} />}
-                  />
-                  <Route path="*" element={<FourZeroFourPage show />} />
-                </Route>
-                <Route
-                  path="/"
-                  element={<LandingPage show={key === pathname && pathname === "/"} />}
-                />
-                <Route path="*" element={<Navigate to="/" />} />
-              </Routes>
-            </ErrorBoundary>
-          </a.div>
-        ))}
-      </Container>
-    </Box>
+                  <Route path="*" element={<Navigate to="/" />} />
+                </Routes>
+              </ErrorBoundary>
+            </a.div>
+          ))}
+        </Container>
+      </Box>
+    </Stack>
   )
 }
